refactor(session): split work item store responsibilities
